+++ lib/appium_lib_core/common/base.rb
@@ -1,7 +1,7 @@
 # The following files have selenium-webdriver related stuff.
 require_relative 'base/driver'
 require_relative 'base/bridge'
-require_relative 'base/bridge/msjsonwp'
+require_relative 'base/bridge/mjsonwp'
 require_relative 'base/bridge/w3c'
 require_relative 'base/capabilities'
 require_relative 'base/http_default'